InnerSource in Action
A first attempt at the InnerSource in Action page. https://github.com/InnerSourceCommons/innersourcecommons.net/issues/22

I added the logos in an image at the top and the quotes are structured on 3 columns underneath.

I think this should be ok for the first pass at the page however we need to figure out a solution for adding more logos to the page.