Create community.md
First draft of the community page, incorporating most of the InnerSource 101 doc. 
Aim of the page is to get people to take the next step after passively consuming content to join the conversation. It will also give an overview of the working groups. 
The images need replacing / upgrading (they're just copied from the front page) - but can work for MVP as is. 
This page should be moved under the Community section as the index page. I couldn't see how to create a folder - so I will create a separate issue to do this and leave that to @voborgus :)