Update _index.md
Creating an index page to introduce all the various learning resources pages. Note that the images will need updating - but this is to create the overall format and the text. the Articles page has been removed to be considered at a later date in line with the latest discussions on that.