fix: Social image default with page override
Pages did not always have a social image unless one was explicitly
set. Since there was no requirement for an image to be set, this
resulted in many pages not having a nice graphic when shared.

This fix creates a side defaultImage variable and sets it. All
subequent pages will us that unless a page-level image is set.
This ensures backwards compatibility and a nice image going
forward.

fixes #171