Update bylaws to the most current set
The bylaws on the ISC website were correctly identified as being out of date. This change switches to the latest approved set dating from February 2020.

It replaces the existing documented bylaws dating from Dec 2017. The outdated bylaws may be an artifact from the old website when it was updated in early 2021. 

This addresses the issue identified by @spier in https://github.com/InnerSourceCommons/innersourcecommons.org/issues/113