Devenv: Run locally with docker
This commit introduces a docker-compose.yml file so that
contributors can run the hugo server locally without having
to install hugo, as long as they have Docker installed (there may
be also compatible alternatives I am not aware of).

I have used the image https://hub.docker.com/r/klakegg/hugo/#!
and kept the base os and version as close as possible
to what is used in the deployment github actions.

Unfortunately this introduces the need to maintain the same
information in both places.