feat(events): Blog post for events
This commit adds the alternative to provide a link to a blog
post related to an event, when the recording is not available.

As the space in the UI cannot currently accomodate an additional
button, it is either a recording, or a blog post.

The feature is used in the community call from 2024-02-27.